requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.9.2
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.9
